Refactor code into multiple classes (part 1)
