+++ .gitignore
@@ -1,1 +1,5 @@
 .Patcher-Temp/* 
+*/__pycache__/*
+build/*.py
+build/compile/*
+flash-patcher.zip
