--- Makefile
@@ -1,8 +1,0 @@
-install-lite:
-	sudo cp flash-patcher.py /usr/local/bin
-	sudo mv /usr/local/bin/flash-patcher.py /usr/local/bin/flash-patcher
-
-install:
-	sudo cp flash-patcher.py /usr/local/bin
-	sudo mv /usr/local/bin/flash-patcher.py /usr/local/bin/flash-patcher
-	sudo cp flash-patcher.json /var/gh-update/flash-patcher.json
