+++ src/compile/__init__.py
@@ -1,0 +1,1 @@
+# empty file, only needed for the zip to detect this folder
